separate file for gaussian distribution function
